Change map loading sequence
NOTE: There's an issue with leaflet and the coexistence with mapbox (Line 90 of HomeMap.tsx when trying to load both maps). This is not an issue of the layers not being triggered. If the leaflet map hasn't been previously loaded, then the map will behave correctly; however, if the leaflet map has been loaded, then the polygons won't be drawn... idk why